could not for the life of me figure out why bombs would never update correctly on the uart. also SysTick_Handler doesnt interrupt correctly.

+++ Menu.h
@@ -6,4 +6,4 @@
 void menuInit(Camera* c);
 void playMenu(void);
 uint8_t difficultyMenu(Camera c);
-void deathMenu(uint16_t score);
+void deathMenu(uint8_t score);
