+++ Menu2.h
@@ -4,5 +4,5 @@
 
 void menuInit2(void);
 void startMenu(void);
-void drawScore(uint16_t score);
+void drawScore(uint8_t score);
 void mBeat(void);
